ADD: create modals for prevent

+++ client/src/pages/Project/Project.js
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { Container, TitleContainer, Wrapper } from "./styles";
+import { Container, MainContainer, TitleContainer, Wrapper } from "./styles";
 import { useAuth } from "../../context/auth-context";
 import { useNavigate, useParams } from "react-router-dom";
 import { listActivities } from "../../services/activity-service";
@@ -9,22 +9,31 @@ import { BiArrowBack } from "react-icons/bi";
 import { StyleBiTrashAlt } from "../Workspace/styles";
 import { deleteProject } from "../../services/project-service";
 import { formatDateToString } from "../../utils";
+import Modal from "../../components/Modal";
 
 export default function Project(){
   const { currentProject, updateListActivites } = useAuth()
   const [ activities, setActivities ] = useState(null)
   const [ parentId, setParentId ] = useState(null)
   const [ editProjectPermission, setEditProjectPermission ] = useState("view")
+  const [showModalDeleteProject, setShowModalDeleteProject] = useState(false)
   const {id} = useParams()
   const nav = useNavigate()
   const handleDeleteProject = () => {
+    const isDeleted = sessionStorage.getItem("isDeleted")
+    console.log("ISDEL", isDeleted)
+    console.log(typeof(isDeleted))
+    if(isDeleted==="true") return setShowModalDeleteProject(true)
+    
     deleteProject(currentProject.id).then(res => {
       console.log(res)
     }).catch(err => {
       console.log(err)
     })
+    sessionStorage.removeItem("isDeleted")
     nav("/workspaces")
   }
+
   useEffect(() => {
     listActivities(id).then(res => {
       console.log(res)
@@ -38,40 +47,61 @@ export default function Project(){
   
   return(
     <Wrapper>
-      <TitleContainer>
-        <div>
-          <BiArrowBack style={{scale:"1.5", cursor:"pointer"}} onClick={() =>{
-            sessionStorage.removeItem("currentProject")
-            nav("/workspaces")
-          }}/>
-          <h2>{currentProject.title}</h2>
-        </div>
-        <div>
-          <p>{formatDateToString(currentProject.init_date)}</p>
-          <p> : </p>
-          <p>{formatDateToString(currentProject.end_date)}</p>
-        </div>
-        <div>
-          <h2>{(currentProject.total_progress*100).toFixed(2)}%</h2>
+      <Modal 
+      showModal={showModalDeleteProject} 
+      setShowModal={setShowModalDeleteProject}
+      onSubmit={()=>{
+        setShowModalDeleteProject(false)
+        sessionStorage.removeItem("isDeleted")
+        deleteProject(currentProject.id).then(res => {
+          console.log(res)
+        }).catch(err => {
+          console.log(err)
+        })
+        nav("/workspaces")
+      }}
+      title="¿Estas seguro que deseas eliminar este proyecto?"
+      text="Eliminar"
+      typeButton="solid"
+      onClick={() => setShowModalDeleteProject(false)}
+      />
+      <MainContainer>
+        <TitleContainer>
+          <div>
+            <BiArrowBack style={{scale:"1.5", cursor:"pointer"}} onClick={() =>{
+              sessionStorage.removeItem("currentProject")
+              
+              nav("/workspaces")
+            }}/>
+            <h2>{currentProject.title}</h2>
+          </div>
+          <div>
+            <p>{formatDateToString(currentProject.init_date)}</p>
+            <p> : </p>
+            <p>{formatDateToString(currentProject.end_date)}</p>
+          </div>
+          <div>
+            <h2>{(currentProject.total_progress*100).toFixed(2)}%</h2>
+            {
+              editProjectPermission !== "view" &&
+              <StyleBiTrashAlt onClick={handleDeleteProject}/>
+            }      
+          </div>
+        </TitleContainer>
+        <Container>
+          {activities && activities.map((activity) => {
+            return(
+              <ActivityCard key={activity.id} activity={activity} editProjectPermission={editProjectPermission}/>
+              )
+            }
+            )}
+            {console.log("PARENTID", parentId)}
           {
             editProjectPermission !== "view" &&
-            <StyleBiTrashAlt onClick={handleDeleteProject}/>
-          }      
-        </div>
-      </TitleContainer>
-      <Container>
-        {activities && activities.map((activity) => {
-          return(
-            <ActivityCard key={activity.id} activity={activity} editProjectPermission={editProjectPermission}/>
-            )
+            <EmptyActivity parent={id}/>
           }
-          )}
-          {console.log("PARENTID", parentId)}
-        {
-          editProjectPermission !== "view" &&
-          <EmptyActivity parent={id}/>
-        }
-      </Container>
+        </Container>
+      </MainContainer>
     </Wrapper>
   )
 }
